refactor: restructure test db setup
