+++ example/src/tests/unit/common.ts
@@ -1,9 +1,6 @@
 import { Chance } from 'chance'
-import {
-  NitroSQLiteConnection,
-  enableSimpleNullHandling,
-} from 'react-native-nitro-sqlite'
-import { testDb as testDbInternal, resetTestDb } from '../db'
+import { enableSimpleNullHandling } from 'react-native-nitro-sqlite'
+import { resetTestDb } from '../db'
 import chai from 'chai'
 
 export function isError(e: unknown): e is Error {
@@ -13,23 +10,7 @@ export function isError(e: unknown): e is Error {
 export const expect = chai.expect
 export const chance = new Chance()
 
-export let testDb: NitroSQLiteConnection
-
 export function setupTestDb() {
   enableSimpleNullHandling(false)
-
-  try {
-    resetTestDb()
-
-    if (testDbInternal == null) throw new Error('Failed to reset test database')
-
-    testDbInternal.execute('DROP TABLE IF EXISTS User;')
-    testDbInternal.execute(
-      'CREATE TABLE User ( id REAL PRIMARY KEY, name TEXT NOT NULL, age REAL, networth REAL) STRICT;',
-    )
-
-    testDb = testDbInternal!
-  } catch (e) {
-    console.warn('Error resetting user database', e)
-  }
+  resetTestDb()
 }
