+++ example/src/tests/unit/specs/execute.spec.ts
@@ -1,9 +1,10 @@
-import { chance, expect, isError, testDb } from '../common'
+import { chance, expect, isError } from '../common'
 import {
   enableSimpleNullHandling,
   NITRO_SQLITE_NULL,
 } from 'react-native-nitro-sqlite'
 import { describe, it } from '../../MochaRNAdapter'
+import { testDb } from '../../db'
 
 export default function registerExecuteUnitTests() {
   describe('execute', () => {
